requirements.txt
torch
torchvision
matplotlib
Pillow
opencv-python
ultralytics
web3
pandas
scipy
flask
flask-cors
werkzeug
